refactor(community): News & Event Section 컴포넌트 CommunityPage 컴포넌트로부터 분리 (#66)
