+++ src/pages/Community/CommunityPage.tsx
@@ -1,75 +1,15 @@
-import axios from 'axios';
 import { JSX } from 'react/jsx-runtime';
-import { useEffect, useRef, useState } from 'react';
-
-import { NewsEventCardInfo } from './Community.types';
 
 import NoticeSection from './Notice/NoticeSection';
-import CategoryCommunity from '@components/CategoryCommunity/CategoryCommunity';
-import NewsEventCard from './NewsEvent/NewsEventCard';
+import NewsEventSection from './NewsEvent/NewsEventSection';
 
 import * as S from './CommunityPage.styled';
 
 const CommunityPage = (): JSX.Element => {
-  const [newsEventCardInfos, setNewsEventCardInfos] = useState<
-    NewsEventCardInfo[]
-  >([]);
-
-  const newsEventTopRef = useRef<HTMLDivElement | null>(null);
-
-  const NewsEventCategory = ['All', 'Recruit', 'Award', 'Alumni'];
-
-  useEffect(() => {
-    const fetchNewsEventCardData = async () => {
-      try {
-        const response = await axios.get('/data/newsEvent.json');
-        const newsEventCardData = response.data.data;
-        console.log(newsEventCardData);
-
-        setNewsEventCardInfos(newsEventCardData);
-      } catch (error) {
-        console.error('News & Event card data fetching error', error);
-      }
-    };
-
-    fetchNewsEventCardData();
-  }, []);
-
   return (
     <S.CommunityPageContainer>
       <NoticeSection />
-
-      {/* News & Event Section */}
-      <S.NewsEventWrapper ref={newsEventTopRef}>
-        <S.NewsEventCategoryContainer>
-          <CategoryCommunity
-            categoryList={NewsEventCategory}
-            ScrollToTopRef={newsEventTopRef}
-          />
-        </S.NewsEventCategoryContainer>
-        <S.NewsEventContainer>
-          <S.NewsEventTitleSection>
-            <S.NewsEventTitle>
-              News & Event <span className="dot">.</span>
-            </S.NewsEventTitle>
-            <S.NewsEventCardGridContainer>
-              <S.BoldDivider />
-              <S.NewsEventCardGrid>
-                {newsEventCardInfos.map((card) => (
-                  <NewsEventCard
-                    key={card.id}
-                    route={`/news-event/${card.id}`}
-                    imageURL={card.imgURL}
-                    deadline={card.deadline}
-                    title={card.title}
-                    category={card.category}
-                  />
-                ))}
-              </S.NewsEventCardGrid>
-            </S.NewsEventCardGridContainer>
-          </S.NewsEventTitleSection>
-        </S.NewsEventContainer>
-      </S.NewsEventWrapper>
+      <NewsEventSection />
     </S.CommunityPageContainer>
   );
 };
