+++ src/pages/Community/NewsEvent/Card/NewsEventCard.styled.ts
