+++ src/styles/global/_normalize.scss
@@ -58,11 +58,6 @@ dd {
   margin: 0;
 }
 
-span {
-  user-select: none;
-  -webkit-user-select: none;
-}
-
 // 5. 리스트 스타일 정리
 ul,
 ol {
